Set additional tracing default tags
I've been looking into why we can't search for custom spans, and my
suspicion is that they're not properly retained if they don't have all
the tags set. Indeed we can search for local spans (env:development),
but searching for production spans (env:production) only returns errored
ones, which are always retained.

The docs are a bit vague about this, but they do mention that they try
to retain a diversity of spans based on the env, service, operation, and
resource tags. We have an operation name set, but not a operation tag,
so maybe that's it? While I'm at it, I'm also adding other tags that are
present in auto-instumented spans.