Tag Elasticsearch as the serivce
This follows our existing convention of using the service name for
library calls. It's not a big difference right now because only hats
uses ES, but if we were to add ES to other services, we would want their
library calls to be attributed to them, and not a generic
"elasticsearch" service.