This time actually use the path group in the resource name
This is fixing up the previous commit, because upon closer inspection it
turns out that there is no path_group tag on these spans, because that
tag is generated later on in the Datadog Agent. We can generate this tag
ourselves though, at least roughly, so this should fix the issue.