feat: Propagate span errors to the service entry span as well
This is required for error tracking to pick them up.

If we already set a a different error on that span, that would likely
happeng later, and overwrite this error, meaning this change only adds
errors where previously we didn't have any.

The main use case for this is the 502s we return for invalid cards,
without throwing an error on the top-level.